backend/requirements.txt
flask>=2.3.0
flask-cors>=3.0.10
google-auth==2.3.3
requests==2.26.0
apscheduler==3.8.1
python-dotenv==0.19.0
SQLAlchemy==1.4.23
firebase-admin==5.0.0
